Rename checkout/charge/sub to have 'connect' in the name for easily understanding intended usage.
