+++ app/app/(payments)/checkout/[id]/page.tsx
@@ -1,4 +1,4 @@
-import { getCheckoutData } from "@/app/services/billing/checkout-service";
+import { getCheckoutData } from "@/app/services/billing/connect-checkout-service";
 import { getSubdomainFromString } from "@/app/services/site/domain-request-service";
 import { TierNotAvailable } from "@/components/tiers/tier-not-available";
 import { headers } from "next/headers";
