+++ app/app/(payments)/checkout/[id]/product-info.tsx
@@ -1,6 +1,6 @@
 import { CHECKOUT_CURRENCY, CHECKOUT_CURRENCY_SYMBOL } from "@/app/config/checkout";
 import { Tier } from "@/app/generated/prisma";
-import { getShortenedCadence } from "@/app/services/billing/checkout-service";
+import { getShortenedCadence } from "@/app/services/billing/connect-checkout-service";
 import { TierDescriptionFeatures } from "@/components/tiers/tier-description-features";
 import { Separator } from "@/components/ui/separator";
 import { parseTierDescription } from "@/lib/utils";
