+++ app/services/tier/tier-version-service.ts
@@ -2,7 +2,7 @@
 
 import { Tier, TierVersion } from "@/app/generated/prisma";
 import prisma from "@/lib/prisma";
-import { checkTierHasSubscribers } from "../billing/subscription-service";
+import { checkTierHasSubscribers } from "../billing/connect-subscription-service";
 import { deactivateStripePrice, type SubscriptionCadence } from "../stripe/stripe-price-service";
 
 export interface VersionContext {
